+++ packages/core/src/actions/clearSuggestedActions.js
@@ -1,0 +1,5 @@
+import setSuggestedActions from './setSuggestedActions';
+
+const clearSuggestedActions = () => setSuggestedActions();
+
+export default clearSuggestedActions;