+++ CHANGELOG.md
@@ -43,7 +43,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
    - Removed [`redux-promise-middleware`](https://www.npmjs.com/package/redux-promise-middleware)
 
 ### Fixed
--  Fix [#2160](https://github.com/microsoft/BotFramework-WebChat/issues/2160). Clear suggested-actions after button click, by [@tdurnford](https://github.com/tdurnford) in PR [#2190](https://github.com/microsoft/BotFramework-WebChat/pull/2190)
+-  Fix [#2160](https://github.com/microsoft/BotFramework-WebChat/issues/2160). Clear suggested actions after clicking on a suggested actions of type `openUrl`, by [@tdurnford](https://github.com/tdurnford) in PR [#2190](https://github.com/microsoft/BotFramework-WebChat/pull/2190)
 
 ## [4.5.0] - 2019-07-10